+++ server/src/test/resources/kmp-module/src/jsMain/kotlin/com/example/common/JsPlatform.kt
@@ -1,0 +1,3 @@
+package com.example.common
+
+actual fun platformName(): String = "JS"